Rename easy_dr -> hertta, make library crate, rename modules
