+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "easy_dr"
+name = "hertta"
 version = "0.1.0"
 edition = "2021"
