+++ src/main.rs
@@ -4,15 +4,12 @@ use std::collections::HashMap;
 use std::env;
 mod main_loop;
 mod utilities;
-mod juliainterface;
-mod structures;
+use hertta::predicer;
 
-use crate::structures::data;
-//use crate::input_data::functions;
 
 
 pub fn start_sending(tx: Sender<main_loop::_Task>) {
-    
+
     tx.send(main_loop::_Task::StartProcess).unwrap();
     tx.send(main_loop::_Task::QuitProcess).unwrap();
     tx.send(main_loop::_Task::StartProcess).unwrap();
@@ -32,7 +29,7 @@ fn main() {
         ("Data2".to_string(), "Value2".to_string()),
     ];
 
-    let time_series1 = data::TimeSeries {
+    let time_series1 = predicer::TimeSeries {
         scenario: "Scenario1".to_string(),
         series: series1,
     };
@@ -41,41 +38,41 @@ fn main() {
         ("Data3".to_string(), "Value3".to_string()),
         ("Data4".to_string(), "Value4".to_string()),
     ];
-    let time_series2 = data::TimeSeries {
+    let time_series2 = predicer::TimeSeries {
         scenario: "Scenario2".to_string(),
         series: series2,
     };
 
     // Step 2: Create a Vec<TimeSeries> containing the created TimeSeries instances
-    let time_series_data_vec: Vec<data::TimeSeries> = vec![time_series1, time_series2];
+    let time_series_data_vec: Vec<predicer::TimeSeries> = vec![time_series1, time_series2];
 
     // Step 3: Create a new TimeSeriesData instance with the Vec<TimeSeries>
-    let time_series_data: data::TimeSeriesData = data::TimeSeriesData {
+    let time_series_data: predicer::TimeSeriesData = predicer::TimeSeriesData {
         ts_data: time_series_data_vec,
     };
 
     //Creating node_diffusion
 
-    let diffusion_1 = data::NodeDiffusion {
+    let diffusion_1 = predicer::NodeDiffusion {
 
         name: String::from("diffusion_1"),
         node1: String::from("interiorair"),
         node2: String::from("buildingenvelope"),
-        diff_coeff: 0.5, 
-        
+        diff_coeff: 0.5,
+
     };
 
-    let diffusion_2 = data::NodeDiffusion {
+    let diffusion_2 = predicer::NodeDiffusion {
 
         name: String::from("diffusion_2"),
         node1: String::from("buildingenvelope"),
         node2: String::from("outside"),
-        diff_coeff: 0.4,   
+        diff_coeff: 0.4,
     };
 
     //Creating node_delay
 
-    let delay_1 = data::NodeDelay {
+    let delay_1 = predicer::NodeDelay {
 
         name: String::from("delay_1"),
         node1: String::from("dh1"),
@@ -83,12 +80,12 @@ fn main() {
         delay: 2.0,
         min_flow: 0.0,
         max_flow: 20.0,
-        
+
     };
 
     //Creating nodes
 
-    let _interiorair = data::Node {
+    let _interiorair = predicer::Node {
         name: String::from("interiorair"),
         is_commodity: false,
         is_state: true,
@@ -99,7 +96,7 @@ fn main() {
         inflow: &time_series_data,
     };
 
-    let _building_envelope = data::Node {
+    let _building_envelope = predicer::Node {
         name: String::from("buildingenvelope"),
         is_commodity: false,
         is_state: true,
@@ -110,7 +107,7 @@ fn main() {
         inflow: &time_series_data,
     };
 
-    let _outside = data::Node {
+    let _outside = predicer::Node {
         name: String::from("outside"),
         is_commodity: false,
         is_state: true,
@@ -121,7 +118,7 @@ fn main() {
         inflow: &time_series_data,
     };
 
-    let _electricitygrid = data::Node {
+    let _electricitygrid = predicer::Node {
         name: String::from("electricitygrid"),
         is_commodity: false,
         is_state: false,
@@ -132,14 +129,14 @@ fn main() {
         inflow: &time_series_data,
     };
 
-    let _node_history_1 = data::NodeHistory {
+    let _node_history_1 = predicer::NodeHistory {
         node: String::from("electricitygrid"),
         steps: &time_series_data,
     };
 
-    let mut _nodes: HashMap<&String, &data::Node> = HashMap::new();
-    let mut _node_diffusion: HashMap<&String, &data::NodeDiffusion> = HashMap::new();
-    let mut _node_delay: HashMap<&String, &data::NodeDelay> = HashMap::new();
+    let mut _nodes: HashMap<&String, &predicer::Node> = HashMap::new();
+    let mut _node_diffusion: HashMap<&String, &predicer::NodeDiffusion> = HashMap::new();
+    let mut _node_delay: HashMap<&String, &predicer::NodeDelay> = HashMap::new();
 
     _nodes.insert(&_interiorair.name, &_interiorair);
     _nodes.insert(&_building_envelope.name, &_building_envelope);
@@ -152,11 +149,11 @@ fn main() {
     _node_delay.insert(&delay_1.name, &delay_1);
 
 
-    let mut _processes: HashMap<&String, &data::Process> = HashMap::new();
+    let mut _processes: HashMap<&String, &predicer::Process> = HashMap::new();
 
     //Creating topology for processes
 
-    let topology1 = data::Topology {
+    let topology1 = predicer::Topology {
 
         source: String::from("electricitygrid"),
         sink: String::from("electricheater"),
@@ -166,7 +163,7 @@ fn main() {
         ramp_down: 1.0
     };
 
-    let topology2 = data::Topology {
+    let topology2 = predicer::Topology {
 
         source: String::from("electricheater"),
         sink: String::from("interiorair"),
@@ -176,11 +173,11 @@ fn main() {
         ramp_down: 1.0
     };
 
-    let topo_vec: Vec<data::Topology> = vec![
+    let topo_vec: Vec<predicer::Topology> = vec![
         topology1,
         topology2,
     ];
-    
+
     //Creating process
 
     //Mitä eff_ops sisältää?
@@ -189,7 +186,7 @@ fn main() {
         ("eff_ops".to_string()),
     ];
 
-    let _electricheater1 = data::Process {
+    let _electricheater1 = predicer::Process {
         name: String::from("electricheater1"),
         is_cf: false,
         is_cf_fix: false,
@@ -213,11 +210,11 @@ fn main() {
     _processes.insert(&_electricheater1.name, &_electricheater1);
 
 
-    let mut _markets: HashMap<&String, &data::Market> = HashMap::new();
-    let mut _groups: HashMap<&String, &data::Group> = HashMap::new();
-    let mut _genconstraints: HashMap<&String, &data::GenConstraint> = HashMap::new();
-    
-    let _npe = data::Market {
+    let mut _markets: HashMap<&String, &predicer::Market> = HashMap::new();
+    let mut _groups: HashMap<&String, &predicer::Group> = HashMap::new();
+    let mut _genconstraints: HashMap<&String, &predicer::GenConstraint> = HashMap::new();
+
+    let _npe = predicer::Market {
         name: String::from("npe"),
         m_type: String::from("energy"),
         node: String::from("electricitygrid"),
@@ -234,49 +231,42 @@ fn main() {
 
     _markets.insert(&_npe.name, &_npe);
 
-    let _p1 = data::Group {
+    let _p1 = predicer::Group {
         name: String::from("p1"),
         g_type: String::from("process"),
         entity: String::from("electricheater")
     };
 
-    /*Key: c_interiorair_up, 
-    Value: GenConstraint("c_interiorair_up", "st", true, 1000.0, 
-    ConFactor[ConFactor("state", ("interiorair", ""), 
-    Predicer.TimeSeriesData(TimeSeries[TimeSeries("s1", Tuple{AbstractString, Number}[("2022-04-20T00:00:00+00:00", 298.15), ("2022-04-20T01:00:00+00:00", 298.15), ("2022-04-20T02:00:00+00:00", 298.15), ("2022-04-20T03:00:00+00:00", 298.15), ("2022-04-20T04:00:00+00:00", 298.15), ("2022-04-20T05:00:00+00:00", 298.15), ("2022-04-20T06:00:00+00:00", 298.15), ("2022-04-20T07:00:00+00:00", 298.15), ("2022-04-20T08:00:00+00:00", 298.15), ("2022-04-20T09:00:00+00:00", 298.15)]), 
+    /*Key: c_interiorair_up,
+    Value: GenConstraint("c_interiorair_up", "st", true, 1000.0,
+    ConFactor[ConFactor("state", ("interiorair", ""),
+    Predicer.TimeSeriesData(TimeSeries[TimeSeries("s1", Tuple{AbstractString, Number}[("2022-04-20T00:00:00+00:00", 298.15), ("2022-04-20T01:00:00+00:00", 298.15), ("2022-04-20T02:00:00+00:00", 298.15), ("2022-04-20T03:00:00+00:00", 298.15), ("2022-04-20T04:00:00+00:00", 298.15), ("2022-04-20T05:00:00+00:00", 298.15), ("2022-04-20T06:00:00+00:00", 298.15), ("2022-04-20T07:00:00+00:00", 298.15), ("2022-04-20T08:00:00+00:00", 298.15), ("2022-04-20T09:00:00+00:00", 298.15)]),
 TimeSeries("s2", Tuple{AbstractString, Number}[("2022-04-20T00:00:00+00:00", 298.15), ("2022-04-20T01:00:00+00:00", 298.15), ("2022-04-20T02:00:00+00:00", 298.15), ("2022-04-20T03:00:00+00:00", 298.15), ("2022-04-20T04:00:00+00:00", 298.15), ("2022-04-20T05:00:00+00:00", 298.15), ("2022-04-20T06:00:00+00:00", 298.15), ("2022-04-20T07:00:00+00:00", 298.15), ("2022-04-20T08:00:00+00:00", 298.15), ("2022-04-20T09:00:00+00:00", 298.15)])]))], Predicer.TimeSeriesData(TimeSeries[])) */
 
     _groups.insert(&_p1.name, &_p1);
 
-    let _confactor = data::ConFactor{
+    let _confactor = predicer::ConFactor{
 
         var_type: String::from(""),
         flow: (String::from(""),String::from("")),
-        data: &time_series_data    
+        data: &time_series_data
     };
 
-    let genconstraint_vec: Vec<data::ConFactor> = vec![
+    let genconstraint_vec: Vec<predicer::ConFactor> = vec![
         _confactor,
     ];
 
-    let _c_interiorair_up = data::GenConstraint {
+    let _c_interiorair_up = predicer::GenConstraint {
 
         name: String::from("c_interiorair_up"),
         gc_type: String::from("st"),
         is_setpoint: true,
         penalty: 1000.0,
         factors: genconstraint_vec,
-        constant: &time_series_data,        
+        constant: &time_series_data,
     };
 
     _genconstraints.insert(&_c_interiorair_up.name, &_c_interiorair_up);
 
-    data::_predicer(false, false, false, false, false, false, true, _nodes, _processes, _markets, _groups, _genconstraints, _node_diffusion, _node_delay, predicer_dir);
-
-
-
-  
-
-
-
+    predicer::_predicer(false, false, false, false, false, false, true, _nodes, _processes, _markets, _groups, _genconstraints, _node_diffusion, _node_delay, predicer_dir);
 }
